WIP: refactor module structure
