+++ examples/getting_started0/lib.rs
@@ -1,6 +1,5 @@
 use ranim::{
-    anims::fading::FadingAnim, color::palettes::manim, items::vitem::geometry::Square,
-    prelude::*,
+    anims::fading::FadingAnim, color::palettes::manim, items::vitem::geometry::Square, prelude::*,
 };
 
 // ANCHOR: construct
