+++ examples/selective_sort/lib.rs
@@ -1,9 +1,8 @@
 use glam::{DVec3, dvec2};
 use rand::{SeedableRng, seq::SliceRandom};
 use ranim::{
-    anims::transform::TransformAnim, color::palettes::manim,
-    items::vitem::geometry::Rectangle, prelude::*, timeline::TimeMark,
-    utils::rate_functions::linear,
+    anims::transform::TransformAnim, color::palettes::manim, items::vitem::geometry::Rectangle,
+    prelude::*, timeline::TimeMark, utils::rate_functions::linear,
 };
 
 fn selective_sort(r: &mut RanimScene, num: usize) {
