+++ packages/ranim-core/src/color/mod.rs
@@ -8,7 +8,7 @@ pub use ::color::HueDirection;
 /// Color preludes
 pub mod prelude {
     pub use super::{rgb, rgb8, rgba, rgba8};
-    pub use crate::{color_macro as color, try_color_macro as try_color};
+    pub use super::{color, try_color};
 }
 
 /// Construct an [`AlphaColor<Srgb>`] from rgb u8, the alpha value will be 255
@@ -31,26 +31,18 @@ pub const fn rgba(r: f32, g: f32, b: f32, a: f32) -> AlphaColor<Srgb> {
     AlphaColor::new([r, g, b, a])
 }
 
-/// A macro to parse color string to [`AlphaColor<Srgb>`]
+/// Parse color string to [`AlphaColor<Srgb>`]
 ///
 /// In its inner it uses [`color::parse_color`]
-#[macro_export]
-macro_rules! color_macro {
-    ($color_str:expr) => {{
-        use ::color::{Srgb, parse_color};
-        parse_color($color_str)
-            .expect("Invalid color string")
-            .to_alpha_color::<Srgb>()
-    }};
+pub fn color(color_str: &str) -> AlphaColor<Srgb> {
+    parse_color(color_str)
+        .expect("Invalid color string")
+        .to_alpha_color::<Srgb>()
 }
 
-/// A macro to parse color string to [`AlphaColor<Srgb>`]
+/// Parse color string to [`AlphaColor<Srgb>`] without panic
 ///
 /// In its inner it uses [`color::parse_color`]
-#[macro_export]
-macro_rules! try_color_macro {
-    ($color_str:expr) => {{
-        use ::color::{Srgb, parse_color};
-        parse_color($color_str).map(|c| c.to_alpha_color::<Srgb>())
-    }};
+pub fn try_color(color_str: &str) -> Result<AlphaColor<Srgb>, ParseError> {
+    parse_color(color_str).map(|c| c.to_alpha_color::<Srgb>())
 }
