+++ packages/ranim-core/src/primitives/mod.rs
@@ -1,4 +1,3 @@
-
 use crate::{
     Extract,
     primitives::{camera_frame::CameraFrame, vitem::VItemPrimitive},
