+++ src/render/pipelines/map_3d_to_2d.rs
@@ -1,9 +1,6 @@
 use std::ops::Deref;
 
-use crate::{
-    render::{CameraUniforms, RenderResource},
-    utils::wgpu::WgpuContext,
-};
+use crate::render::{CameraUniforms, RenderResource, WgpuContext};
 
 pub struct Map3dTo2dPipeline {
     pipeline: wgpu::ComputePipeline,
